requirements.student2.txt
elasticsearch>=8,<9
python-dotenv
